refactor(slots): investigate drift-delete, land drift.py kept (P3-slots §1c)
Spec §6 asked to attempt deleting compute_config_drift on the hypothesis
that "running argv equals rendered argv by construction" once §11.2
PortAuthority issues --model/--alias/port authoritatively. That seam does
not exist yet (plan §23.2: "note the interface; don't build it"), and
tests/slots/test_config_drift_aliases.py::test_real_drift_still_detected_across_spellings
proves a genuine divergence (-b 2048 rendered vs --batch-size 512 running)
is still correctly flagged today — i.e. a container CAN run stale argv
after a TOML edit without a restart, which is exactly the class of bug
this comparator exists to catch. Decision: KEEP, do not delete; extract
verbatim to slots/drift.py instead.

Moves: compute_config_drift, _argv_values, _resolve_drift_flags,
_config_drift_values_equal, _CONFIG_DRIFT_KEYS. compute_config_drift takes
a narrow DriftHost protocol (_is_active/_maybe_load_config/
_resolve_model_info). SlotManager.compute_config_drift is now a one-line
delegator; status()'s include_config_drift param and
api/routes/updater.py's drift-preview route are unchanged (both just call
the still-present method).

Re-exports from manager.py: _CONFIG_DRIFT_KEYS, _argv_values (both
imported directly by tests/slots/test_config_drift_aliases.py).

No behavior change. tests/slots (379) + tests/api updater/drift (40) green.
